Move some utils to AbstractCommitCoordinatorBuilder
