+++ kernel/kernel-api/src/main/java/io/delta/kernel/coordinatedcommits/CommitCoordinatorUtils.java
@@ -18,19 +18,13 @@ package io.delta.kernel.coordinatedcommits;
 
 import static io.delta.kernel.internal.TableConfig.COORDINATED_COMMITS_COORDINATOR_NAME;
 
-import io.delta.kernel.config.ConfigurationProvider;
 import io.delta.kernel.data.Row;
 import io.delta.kernel.engine.Engine;
 import io.delta.kernel.engine.coordinatedcommits.UpdatedActions;
-import io.delta.kernel.internal.DeltaErrors;
 import io.delta.kernel.internal.fs.Path;
 import io.delta.kernel.utils.CloseableIterator;
 import io.delta.kernel.utils.FileStatus;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.util.Map;
-import java.util.NoSuchElementException;
-import java.util.UUID;
 
 /** Various public utility methods related to Coordinated Commits. */
 public class CommitCoordinatorUtils {
@@ -38,63 +32,6 @@ public class CommitCoordinatorUtils {
   private static final String COMMIT_SUBDIR = "_commits";
 
   private CommitCoordinatorUtils() {}
-
-  /**
-   * Builds the underlying {@link CommitCoordinatorClient} associated with the given commit
-   * coordinator name.
-   *
-   * <ul>
-   *   <li>Determines the specific builder configuration lookup key.
-   *   <li>Grabs the corresponding builder className for that key from the provided sessionConfig.
-   *   <li>Instantiates a new instance of that {@link AbstractCommitCoordinatorBuilder}.
-   *   <li>Invokes the builder's build method, passing along the sessionConfig and commit
-   *       coordinator config map.
-   * </ul>
-   *
-   * @param commitCoordinatorName the commit coordinator name
-   * @param sessionConfig The session-level configuration that may represent environment-specific
-   *     configurations such as {@code HadoopConf} or {@code SparkConf}. This sessionConfig is used
-   *     to look up the right builder className to instantiate for the given commit coordinator
-   *     name. It can also be used by builders to lookup per-session configuration values unique to
-   *     that builder.
-   * @param commitCoordinatorConf the parsed value of the Delta table property {@link
-   *     io.delta.kernel.internal.TableConfig#COORDINATED_COMMITS_TABLE_CONF} and represents the
-   *     configuration properties for describing the Delta table to the commit-coordinator.
-   * @return the {@link CommitCoordinatorClient} corresponding to the given commit coordinator name
-   */
-  public static CommitCoordinatorClient buildCommitCoordinatorClient(
-      String commitCoordinatorName,
-      ConfigurationProvider sessionConfig,
-      Map<String, String> commitCoordinatorConf) {
-    final String builderConfKey = getCommitCoordinatorBuilderConfKey(commitCoordinatorName);
-
-    final String builderClassName;
-    try {
-      builderClassName = sessionConfig.getNonNull(builderConfKey);
-    } catch (NoSuchElementException | IllegalStateException ex) {
-      throw DeltaErrors.unknownCommitCoordinator(commitCoordinatorName, builderConfKey);
-    }
-
-    try {
-      return Class.forName(builderClassName)
-          .asSubclass(AbstractCommitCoordinatorBuilder.class)
-          .getConstructor()
-          .newInstance()
-          .build(sessionConfig, commitCoordinatorConf);
-    } catch (ClassNotFoundException
-        | NoSuchMethodException
-        | InstantiationException
-        | IllegalAccessException
-        | InvocationTargetException e) {
-      throw DeltaErrors.couldNotInstantiateCommitCoordinatorClient(
-          commitCoordinatorName, builderClassName, e);
-    }
-  }
-
-  /** Returns the builder configuration key for the given commit coordinator name */
-  public static String getCommitCoordinatorBuilderConfKey(String ccName) {
-    return String.format("io.delta.kernel.commitCoordinatorBuilder.%s.impl", ccName);
-  }
 
   /** Returns true if the commit is a coordinated commits to filesystem conversion. */
   public static boolean isCoordinatedCommitsToFSConversion(
